Don't install puppetlabs repo when MANAGE_REPOS=false
Currently, p-o-i allways tries to install puppet package
from puppetlabs repo. In some cases, as RDO-CI we use p-o-i
to test puppet modules from RDO distro and it's recommended
to test it using puppet version provided in RDO repos which
uses different paths.

This patch provides a way to override this behavior by using
env variables as:

MANAGE_REPOS=false
PUPPET_PKG=puppet
PUPPETFILE_DIR=/etc/puppet/modules

+++ install_modules.sh
@@ -10,10 +10,12 @@ fi
 # NOTE(aschultz): since puppet 3 is now EOL, and beaker-puppet_install_helper
 # version 0.6.0 has made the agent version the default, we need to symlink
 # puppet to the /opt/puppetlabs version when specifically not version 3.
-export PUPPET_BASE_PATH=/etc/puppetlabs/code
-export PATH=${PATH}:/opt/puppetlabs/bin:/opt/puppetlabs/puppet/bin
 # Workaround to deploy puppet for beaker jobs
-sudo -E ln -sfn /opt/puppetlabs/bin/puppet /usr/sbin/puppet
+if [ -e /opt/puppetlabs/bin/puppet ]; then
+    export PUPPET_BASE_PATH=/etc/puppetlabs/code
+    export PATH=${PATH}:/opt/puppetlabs/bin:/opt/puppetlabs/puppet/bin
+    sudo -E ln -sfn /opt/puppetlabs/bin/puppet /usr/sbin/puppet
+fi
 
 export SCRIPT_DIR=$(cd `dirname $0` && pwd -P)
 export PUPPETFILE_DIR=${PUPPETFILE_DIR:-${PUPPET_BASE_PATH}/modules}
